ct/src/requirements.txt
hyper
xmltodict
pytest-ordering
# parallelism:
pytest-xdist
# parallelism & concurrency:
pytest-parallel
